+++ WaterReservoir.cpp
@@ -1,8 +1,8 @@
 #include "WaterReservoir.h"
 
-WaterReservoir::WaterReservoir(string name, string municipality, unsigned long id, string code, unsigned long maxDelivery)
+WaterReservoir::WaterReservoir(string name, string municipality, double id, string code, double maxDelivery)
     : name(name), municipality(municipality), id(id), code(code), maxDelivery(maxDelivery) {}
 
-unsigned long WaterReservoir::getMaxDelivery() {
+double WaterReservoir::getMaxDelivery() {
     return this->maxDelivery;
 }
